+++ dev-packages/type-check/ts3.8-test/index.ts
@@ -23,6 +23,11 @@ declare global {
   interface PerformanceEntry {}
   interface Performance {}
 }
+
+declare module 'react-native' {
+  export interface TurboModule {}
+}
+
 import 'react-native';
 
 // we need to import the SDK to ensure tsc check the types