+++ packages/core/src/js/tracing/integrations/appStart.ts
@@ -1,5 +1,5 @@
-/* eslint-disable complexity */
-import type { Client, Event, Integration, SpanJSON, TransactionEvent } from '@sentry/core';
+/* eslint-disable complexity, max-lines */
+import type { Client, Event, Integration, Span, SpanJSON, TransactionEvent } from '@sentry/core';
 import {
   getCapturedScopesOnSpan,
   getClient,
@@ -11,13 +11,14 @@ import {
   timestampInSeconds,
 } from '@sentry/core';
 
+import { getAppRegistryIntegration } from '../../integrations/appRegistry';
 import {
   APP_START_COLD as APP_START_COLD_MEASUREMENT,
   APP_START_WARM as APP_START_WARM_MEASUREMENT,
 } from '../../measurements';
 import type { NativeAppStartResponse } from '../../NativeRNSentry';
 import type { ReactNativeClientOptions } from '../../options';
-import { convertSpanToTransaction, setEndTimeValue } from '../../utils/span';
+import { convertSpanToTransaction, isRootSpan, setEndTimeValue } from '../../utils/span';
 import { NATIVE } from '../../wrapper';
 import {
   APP_START_COLD as APP_START_COLD_OP,
@@ -26,6 +27,7 @@ import {
 } from '../ops';
 import { SPAN_ORIGIN_AUTO_APP_START, SPAN_ORIGIN_MANUAL_APP_START } from '../origin';
 import { SEMANTIC_ATTRIBUTE_SENTRY_OP } from '../semanticAttributes';
+import { setMainThreadInfo } from '../span';
 import { createChildSpanJSON, createSpanJSON, getBundleStartTimestampMs } from '../utils';
 
 const INTEGRATION_NAME = 'AppStart';
@@ -136,20 +138,40 @@ export const appStartIntegration = ({
   let _client: Client | undefined = undefined;
   let isEnabled = true;
   let appStartDataFlushed = false;
+  let afterAllSetupCalled = false;
+  let firstStartedActiveRootSpanId: string | undefined = undefined;
 
   const setup = (client: Client): void => {
     _client = client;
-    const clientOptions = client.getOptions() as ReactNativeClientOptions;
+    const { enableAppStartTracking } = client.getOptions() as ReactNativeClientOptions;
 
-    const { enableAppStartTracking } = clientOptions;
     if (!enableAppStartTracking) {
       isEnabled = false;
       logger.warn('[AppStart] App start tracking is disabled.');
     }
+
+    client.on('spanStart', recordFirstStartedActiveRootSpanId);
   };
 
-  const afterAllSetup = (_client: Client): void => {
+  const afterAllSetup = (client: Client): void => {
+    if (afterAllSetupCalled) {
+      return;
+    }
+    afterAllSetupCalled = true;
+
     // TODO: automatically set standalone based on the presence of the native layer navigation integration
+
+    getAppRegistryIntegration(client)?.onRunApplication(() => {
+      if (appStartDataFlushed) {
+        logger.log('[AppStartIntegration] Resetting app start data flushed flag based on runApplication call.');
+        appStartDataFlushed = false;
+        firstStartedActiveRootSpanId = undefined;
+      } else {
+        logger.log(
+          '[AppStartIntegration] Waiting for initial app start was flush, before updating based on runApplication call.',
+        );
+      }
+    });
   };
 
   const processEvent = async (event: Event): Promise<Event> => {
@@ -165,6 +187,27 @@ export const appStartIntegration = ({
     await attachAppStartToTransactionEvent(event as TransactionEvent);
 
     return event;
+  };
+
+  const recordFirstStartedActiveRootSpanId = (rootSpan: Span): void => {
+    if (firstStartedActiveRootSpanId) {
+      return;
+    }
+
+    if (!isRootSpan(rootSpan)) {
+      return;
+    }
+
+    setFirstStartedActiveRootSpanId(rootSpan.spanContext().spanId);
+  };
+
+  /**
+   * For testing purposes only.
+   * @private
+   */
+  const setFirstStartedActiveRootSpanId = (spanId: string | undefined): void => {
+    firstStartedActiveRootSpanId = spanId;
+    logger.debug('[AppStart] First started active root span id recorded.', firstStartedActiveRootSpanId);
   };
 
   async function captureStandaloneAppStart(): Promise<void> {
@@ -208,12 +251,24 @@ export const appStartIntegration = ({
 
   async function attachAppStartToTransactionEvent(event: TransactionEvent): Promise<void> {
     if (appStartDataFlushed) {
-      // App start data is only relevant for the first transaction
+      // App start data is only relevant for the first transaction of the app run
+      return;
+    }
+
+    if (!firstStartedActiveRootSpanId) {
+      logger.warn('[AppStart] No first started active root span id recorded. Can not attach app start.');
       return;
     }
 
     if (!event.contexts?.trace) {
       logger.warn('[AppStart] Transaction event is missing trace context. Can not attach app start.');
+      return;
+    }
+
+    if (firstStartedActiveRootSpanId !== event.contexts.trace.span_id) {
+      logger.warn(
+        '[AppStart] First started active root span id does not match the transaction event span id. Can not attached app start.',
+      );
       return;
     }
 
@@ -252,6 +307,17 @@ export const appStartIntegration = ({
     if (!__DEV__ && appStartDurationMs >= MAX_APP_START_DURATION_MS) {
       // Dev builds can have long app start waiting over minute for the first bundle to be produced
       logger.warn('[AppStart] App start duration is over a minute long, not adding app start span.');
+      return;
+    }
+
+    if (appStartDurationMs < 0) {
+      // This can happen when MainActivity on Android is recreated,
+      // and the app start end timestamp is not updated, for example
+      // due to missing `Sentry.wrap(RootComponent)` call.
+      logger.warn(
+        '[AppStart] Last recorded app start end timestamp is before the app start timestamp.',
+        'This is usually caused by missing `Sentry.wrap(RootComponent)` call.',
+      );
       return;
     }
 
@@ -332,7 +398,8 @@ export const appStartIntegration = ({
     afterAllSetup,
     processEvent,
     captureStandaloneAppStart,
-  };
+    setFirstStartedActiveRootSpanId,
+  } as AppStartIntegration;
 };
 
 function setSpanDurationAsMeasurementOnTransactionEvent(event: TransactionEvent, label: string, span: SpanJSON): void {
@@ -360,19 +427,25 @@ function createJSExecutionStartSpan(
     return undefined;
   }
 
+  const bundleStartTimestampSeconds = bundleStartTimestampMs / 1000;
+  if (bundleStartTimestampSeconds < parentSpan.start_timestamp) {
+    logger.warn('Bundle start timestamp is before the app start span start timestamp. Skipping JS execution span.');
+    return undefined;
+  }
+
   if (!rootComponentCreationTimestampMs) {
     logger.warn('Missing the root component first constructor call timestamp.');
     return createChildSpanJSON(parentSpan, {
       description: 'JS Bundle Execution Start',
-      start_timestamp: bundleStartTimestampMs / 1000,
-      timestamp: bundleStartTimestampMs / 1000,
+      start_timestamp: bundleStartTimestampSeconds,
+      timestamp: bundleStartTimestampSeconds,
       origin: SPAN_ORIGIN_AUTO_APP_START,
     });
   }
 
   return createChildSpanJSON(parentSpan, {
     description: 'JS Bundle Execution Before React Root',
-    start_timestamp: bundleStartTimestampMs / 1000,
+    start_timestamp: bundleStartTimestampSeconds,
     timestamp: rootComponentCreationTimestampMs / 1000,
     origin: isRootComponentCreationTimestampMsManual ? SPAN_ORIGIN_MANUAL_APP_START : SPAN_ORIGIN_AUTO_APP_START,
   });
@@ -382,18 +455,22 @@ function createJSExecutionStartSpan(
  * Adds native spans to the app start span.
  */
 function convertNativeSpansToSpanJSON(parentSpan: SpanJSON, nativeSpans: NativeAppStartResponse['spans']): SpanJSON[] {
-  return nativeSpans.map(span => {
-    if (span.description === 'UIKit init') {
-      return createUIKitSpan(parentSpan, span);
-    }
+  return nativeSpans
+    .filter(span => span.start_timestamp_ms / 1000 >= parentSpan.start_timestamp)
+    .map(span => {
+      if (span.description === 'UIKit init') {
+        return setMainThreadInfo(createUIKitSpan(parentSpan, span));
+      }
 
-    return createChildSpanJSON(parentSpan, {
-      description: span.description,
-      start_timestamp: span.start_timestamp_ms / 1000,
-      timestamp: span.end_timestamp_ms / 1000,
-      origin: SPAN_ORIGIN_AUTO_APP_START,
+      return setMainThreadInfo(
+        createChildSpanJSON(parentSpan, {
+          description: span.description,
+          start_timestamp: span.start_timestamp_ms / 1000,
+          timestamp: span.end_timestamp_ms / 1000,
+          origin: SPAN_ORIGIN_AUTO_APP_START,
+        }),
+      );
     });
-  });
 }
 
 /**