+++ samples/expo/app/_layout.tsx
@@ -11,6 +11,7 @@ import { ErrorEvent } from '@sentry/core';
 import { isExpoGo } from '../utils/isExpoGo';
 import { LogBox } from 'react-native';
 import { isWeb } from '../utils/isWeb';
+import * as ImagePicker from 'expo-image-picker';
 
 export {
   // Catch any errors thrown by the Layout component.
@@ -57,6 +58,9 @@ Sentry.init({
       }),
       navigationIntegration,
       Sentry.reactNativeTracingIntegration(),
+      Sentry.feedbackIntegration({
+        imagePicker: ImagePicker,
+      }),
     );
     if (isWeb()) {
       integrations.push(Sentry.browserReplayIntegration());