+++ samples/react-native/e2e/jest.config.android.js
@@ -1,0 +1,8 @@
+const path = require('path');
+const baseConfig = require('./jest.config.base');
+
+/** @type {import('@jest/types').Config.InitialOptions} */
+module.exports = {
+  ...baseConfig,
+  globalSetup: path.resolve(__dirname, 'setup.android.ts'),
+};